requirements.txt
python-dotenv
jsonpath
httpx